css changes, and implementation all complete

+++ apps/modernization-ui/src/apps/page-builder/page/management/edit/section/Sections.tsx
@@ -1,5 +1,5 @@
 import { PagesQuestion, PagesSection, SectionControllerService } from 'apps/page-builder/generated';
-import React, { RefObject } from 'react';
+import React, { RefObject, useRef } from 'react';
 
 import styles from './section.module.scss';
 import { Section } from './Section';
@@ -7,6 +7,7 @@ import { ModalRef } from '@trussworks/react-uswds';
 import { authorization } from 'authorization';
 import { usePageManagement } from '../../usePageManagement';
 import { useAlert } from 'alert';
+import { StatusModal } from '../../status/StatusModal';
 
 type Props = {
     sections: PagesSection[];
@@ -27,17 +28,23 @@ export const Sections = ({
 }: Props) => {
     const { page, fetch } = usePageManagement();
 
+    const statusModalRef = useRef<ModalRef>(null);
+
     const { showAlert } = useAlert();
 
     const handleDeleteSection = (section: PagesSection) => {
-        SectionControllerService.deleteSectionUsingDelete({
-            authorization: authorization(),
-            page: page.id,
-            sectionId: section.id
-        }).then(() => {
-            showAlert({ message: `You've successfully deleted section!`, type: `success` });
-            fetch(page.id);
-        });
+        if (section.subSections.length > 0) {
+            statusModalRef.current?.toggleModal(undefined, true);
+        } else {
+            SectionControllerService.deleteSectionUsingDelete({
+                authorization: authorization(),
+                page: page.id,
+                sectionId: section.id
+            }).then(() => {
+                showAlert({ message: `You've successfully deleted section!`, type: `success` });
+                fetch(page.id);
+            });
+        }
     };
 
     return (
@@ -55,6 +62,18 @@ export const Sections = ({
                     />
                 </React.Fragment>
             ))}
+            <StatusModal
+                modal={statusModalRef}
+                messageHeader="Section cannot be deleted."
+                title={'Warning'}
+                message={
+                    'This section contains elements (subsections and questions) inside it. Remove the contents first, and then the section can be deleted.'
+                }
+                onConfirm={() => {
+                    statusModalRef.current?.toggleModal(undefined, false);
+                }}
+                confirmText="Okay"
+            />
         </div>
     );
 };
